requirements.txt
web3==6.15.1
requests==2.31.0
python-dotenv==1.0.0
supabase==2.3.4
psycopg2-binary==2.9.9
pandas==2.1.4
streamlit==1.29.0
plotly==5.17.0
schedule==1.2.0